Modified validations, Database name

+++ src/main/java/com/csye6225/CloudAssignment2/Controller/AssignmentController.java
@@ -14,6 +14,7 @@ import org.apache.coyote.Request;
 import org.apache.tomcat.util.http.parser.Authorization;
 import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.CacheControl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -44,26 +45,32 @@ public class AssignmentController {
     @Autowired
     private AccountRepository accountRepository;
     @PostMapping()
-    public ResponseEntity<?> createAssignment(@RequestBody Assignment assignment, Authentication authentication){
-
-//        String authHeader= request.getHeader("Authorization");
-//         boolean result=accountService.validation(authHeader, assignment);
+    public ResponseEntity<?> createAssignment(@RequestBody Assignment assignment){
+        if(request.getQueryString()!= null){
+            return ResponseEntity.status(400).cacheControl(CacheControl.noCache().mustRevalidate()).build();
+        }
 
         try {
             Assignment savedAssignment = assignmentService.saveAssignment(assignment);
-            return ResponseEntity.ok(savedAssignment);
+            return ResponseEntity.status(201).body(savedAssignment);
 
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
     @GetMapping
-    public List<Assignment> getAssignments() {
+    public ResponseEntity<Object> getAssignments(@RequestBody(required = false) Object body) {
+        if(body!=null || request.getQueryString()!= null){
+            return ResponseEntity.status(400).cacheControl(CacheControl.noCache().mustRevalidate()).build();
+        }
 
-        return assignmentService.getAllAssignments();
+        return ResponseEntity.ok(assignmentService.getAllAssignments());
     }
     @GetMapping("/{id}")
-    public ResponseEntity<Assignment> getAssignmentById(@PathVariable UUID id) {
+    public ResponseEntity<Assignment> getAssignmentById(@PathVariable UUID id,@RequestBody(required = false) Object body) {
+        if(body!=null || request.getQueryString()!= null){
+            return ResponseEntity.status(400).cacheControl(CacheControl.noCache().mustRevalidate()).build();
+        }
         Optional<Assignment> assignment = assignmentService.getAssignmentById(id);
 
         if (assignment.isPresent()) {
@@ -74,16 +81,17 @@ public class AssignmentController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateAssignment(@PathVariable UUID id, @RequestBody Assignment updatedAssignment) {
+    public ResponseEntity<Object> updateAssignment(@PathVariable UUID id, @RequestBody(required = false) Assignment updatedAssignment) {
+        if(updatedAssignment==null || request.getQueryString()!= null){
+            return ResponseEntity.status(400).cacheControl(CacheControl.noCache().mustRevalidate()).build();
+        }
         Optional<Assignment> existingAssignment = assignmentService.getAssignmentById(id);
 
         String assignmentName = updatedAssignment.getName();
         if (assignmentName.isEmpty() || assignmentName.isEmpty()||assignmentName.trim().isEmpty()) {
             return ResponseEntity.badRequest().body("The name of Assignment shouldn't be empty");
         }
-        if (assignmentRepository.existsByName(assignmentName)) {
-            return ResponseEntity.badRequest().body("The name of Assignment already exists");
-        }
+
 
         UUID gettingId = (UUID) request.getSession().getAttribute("accountId");
 
@@ -97,11 +105,15 @@ public class AssignmentController {
                 assignment.setNum_of_attempts(updatedAssignment.getNum_of_attempts());
                 assignment.setDeadline(updatedAssignment.getDeadline());
                 assignment.setAssignmentUpdated(Date.valueOf(LocalDate.now()));
-                Assignment updated = assignmentService.saveAssignment(assignment);
-                return ResponseEntity.ok(updated);
+                try{
+                    assignmentService.saveAssignment(assignment);
+                    return ResponseEntity.status(204).build();
+                }catch (Exception ex){
+                    return ResponseEntity.status(400).body(ex.getMessage());
+                }
             }
             else {
-       return  ResponseEntity.status(403).build();
+                return  ResponseEntity.status(403).build();
             }
         } else {
             return ResponseEntity.notFound().build();
@@ -109,14 +121,17 @@ public class AssignmentController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteAssignment(@PathVariable UUID id) {
+    public ResponseEntity<String> deleteAssignment(@PathVariable UUID id,@RequestBody(required = false) Object body) {
+        if(body!=null || request.getQueryString()!= null){
+            return ResponseEntity.status(400).cacheControl(CacheControl.noCache().mustRevalidate()).build();
+        }
         Optional<Assignment> existingAssignment = assignmentService.getAssignmentById(id);
         UUID deleteid = (UUID) request.getSession().getAttribute("accountId");
         if (existingAssignment.isPresent()) {
             Assignment assignment1=existingAssignment.get();
             if(deleteid.equals(assignment1.getCreatedBy().getId())) {
                 assignmentService.deleteAssignment(id);
-                return ResponseEntity.status(200).body("Successfully deleted Assignment with id:" + id);
+                return ResponseEntity.status(204).build();
             }
             else{
                 return ResponseEntity.status(403).build();
